update vari e sistemazione dei mapper

+++ src/main/java/servlets/AdminFunctions.java
@@ -282,11 +282,13 @@ public class AdminFunctions extends HttpServlet {
                         float totalRefund = 0f;
                         float refoundPercentage = 0.8f;
                         for (SeatDetailRequest sdr : seatDetailRequests) {
-                            float ticket_price = sdr.getPrice();
-                            int room_number = showMapper.getRoomNumberFromCode(code);
-                            int id_seat = seatMapper.getIdSeat(room_number, sdr.getS_row(), sdr.getS_column());
-                            userMapper.deletePaymentFromCode(code, id_seat);
-                            totalRefund += ticket_price;
+                            if (sdr.isChecked()) {
+                                float ticket_price = Float.parseFloat(sdr.getPrice());
+                                int room_number = showMapper.getRoomNumberFromCode(code);
+                                int id_seat = seatMapper.getIdSeat(room_number, Integer.parseInt(sdr.getS_row()), Integer.parseInt(sdr.getS_column()));
+                                userMapper.deletePaymentFromCode(code, id_seat);
+                                totalRefund += ticket_price;
+                            }
                         }
                         float computedRefound = totalRefund * refoundPercentage;
                         String username = userMapper.getPaymentFromCode(code).get(0).getUsername();
